requirements.txt
filelock==3.13.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.0
pendulum==3.0.0
pyarrow==14.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
ruamel.yaml==0.18.2
six==1.16.0
sympy==1.12
time-machine==2.14.0
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.10.0
tzdata==2024.1
